Split out base integration test class and compatibility helper
