+++ src/test/groovy/org/gradle/profiler/fixtures/AbstractBaseProfilerIntegrationTest.groovy
@@ -1,0 +1,302 @@
+package org.gradle.profiler.fixtures
+
+import org.gradle.api.JavaVersion
+import org.gradle.profiler.Main
+import org.gradle.util.GradleVersion
+import org.junit.Rule
+import org.junit.rules.TemporaryFolder
+
+import javax.annotation.Nullable
+import java.util.regex.Pattern
+
+import static org.gradle.profiler.fixtures.compatibility.gradle.GradleVersionCompatibility.minimalGradleVersionSupportingJava11
+import static org.gradle.profiler.fixtures.compatibility.gradle.GradleVersionCompatibility.minimalGradleVersionSupportingJava17
+
+abstract class AbstractBaseProfilerIntegrationTest extends AbstractIntegrationTest {
+
+    private static int NUMBER_OF_HEADERS = 4
+
+    static final SAMPLE = "-?\\d+(?:\\.\\d+)"
+
+    @Rule
+    TemporaryFolder tmpDir = new TemporaryFolder()
+
+    File projectDir
+    File outputDir
+
+    LogFile getLogFile() {
+        def f = new File(outputDir, "profile.log")
+        assert f.isFile()
+        return new LogFile(f)
+    }
+
+    ReportFile getResultFile() {
+        def f = new File(outputDir, "benchmark.csv")
+        assert f.isFile()
+
+        return new ReportFile(f)
+    }
+
+    File getBuildFile() {
+        return new File(projectDir, "build.gradle")
+    }
+
+    File getSettingsFile() {
+        return new File(projectDir, "settings.gradle")
+    }
+
+    File file(String path) {
+        return new File(projectDir, path)
+    }
+
+    void run(List<String> moreArgs) {
+        def args = [
+            "--project-dir", projectDir.absolutePath,
+            "--output-dir", outputDir.absolutePath,
+        ] + moreArgs
+
+        new Main().run(*args)
+    }
+
+    // It seems it's not effectful to pass `org.gradle.java.home` as a system property or as part of jvmargs
+    def downgradeDaemonJvmIfTestJvmUnsupported(String gradleVersion, @Nullable File dir = null) {
+        def currentJvm = JavaVersion.current()
+        def targetGradleVersion = GradleVersion.version(gradleVersion)
+
+        String downgradeJavaHome = null
+        if (currentJvm.isCompatibleWith(JavaVersion.VERSION_11) && targetGradleVersion < minimalGradleVersionSupportingJava11) {
+            downgradeJavaHome = getJavaHomeProperty("javaHomes.java8", "Java 8")
+        } else if (currentJvm.isCompatibleWith(JavaVersion.VERSION_17) && targetGradleVersion < minimalGradleVersionSupportingJava17) {
+            downgradeJavaHome = getJavaHomeProperty("javaHomes.java11", "Java 11")
+        }
+
+        if (downgradeJavaHome != null) {
+            def parentDir = dir != null ? dir : projectDir
+            new File(parentDir, "gradle.properties") << "\norg.gradle.java.home=${downgradeJavaHome}"
+        }
+    }
+
+    private static String getJavaHomeProperty(String systemProperty, String description) {
+        def javaHome = System.getProperty(systemProperty)
+        if (javaHome == null) {
+            throw new IllegalStateException("This test requires Java Home for $description to be provided in the '$systemProperty' system property")
+        }
+        return javaHome
+    }
+
+    def setup() {
+        projectDir = tmpDir.newFolder()
+        outputDir = tmpDir.newFolder()
+        settingsFile.text = """
+            rootProject.name = "example-project"
+        """
+    }
+
+    def instrumentedBuildScript() {
+        buildFile.text = """
+apply plugin: BasePlugin
+println "<gradle-version: " + gradle.gradleVersion + ">"
+println "<tasks: " + gradle.startParameter.taskNames + ">"
+println "<daemon: " + gradle.services.get(org.gradle.internal.environment.GradleBuildEnvironment).longLivingProcess + ">"
+println "<invocations: " + (++Counter.invocations) + ">"
+
+class Counter {
+    static int invocations = 0
+}
+"""
+    }
+
+    def writeBuckw() {
+        def buckw = file("buckw")
+        buckw.text = '''
+#!/usr/bin/env bash
+
+echo "[-] PARSING BUCK FILES...FINISHED 0.3s [100%]"
+if [ $1 = "targets" ]
+then
+    if [ "$2" = "--type" ]
+    then
+        echo "//target:$3_1"
+        echo "//target:$3_2"
+        echo "//target/child:$3_3"
+        echo "//target/child:$3_4"
+    else
+        echo "//target:android_binary"
+        echo "//target:java_library"
+        echo "//target:cpp_library"
+        echo "//target/child:android_library"
+        echo "//target/child:cpp_library"
+    fi
+else
+    echo "building $@"
+fi
+'''
+        buckw.executable = true
+    }
+
+    def createSimpleBazelWorkspace() {
+        new File(projectDir, "WORKSPACE").createNewFile()
+        new File(projectDir, "BUILD").text = '''
+genrule(
+  name = "hello",
+  outs = ["hello_world.txt"],
+  cmd = "echo Hello World > $@",
+)'''
+    }
+
+    static class FileFixture {
+        final List<String> lines
+
+        FileFixture(File logFile) {
+            lines = logFile.readLines()
+        }
+
+        @Override
+        String toString() {
+            return lines.join("\n")
+        }
+
+        /**
+         * Contains a single line with the given text.
+         */
+        boolean containsOne(String str) {
+            return find(str).size() == 1
+        }
+
+        /**
+         * Locates the lines containing the given string
+         */
+        List<String> find(String str) {
+            lines.findAll { it.contains(str) }
+        }
+
+        /**
+         * Locates the lines containing the given pattern
+         */
+        List<String> find(Pattern pattern) {
+            lines.findAll { it.matches(pattern) }
+        }
+
+        String getText() {
+            lines.join("\n")
+        }
+    }
+
+    static class ReportFile extends FileFixture {
+        ReportFile(File logFile) {
+            super(logFile)
+        }
+
+        /**
+         * Asserts that this report file contains a single benchmarking scenario that uses a warm daemon.
+         */
+        void containsWarmDaemonScenario(String gradleVersion, String name = "default", List<String> tasks) {
+            def lines = this.lines
+            assert lines.size() == totalLinesForExecutions(16)
+            assert lines.get(0) == "scenario,${name}"
+            assert lines.get(1) == "version,Gradle ${gradleVersion}"
+            assert lines.get(2) == "tasks,${tasks.join(", ")}"
+            assert lines.get(3) == "value,total execution time"
+            assert lines.get(4).matches("warm-up build #1,$SAMPLE")
+            assert lines.get(9).matches("warm-up build #6,$SAMPLE")
+            assert lines.get(10).matches("measured build #1,$SAMPLE")
+            assert lines.get(11).matches("measured build #2,$SAMPLE")
+            assert lines.get(19).matches("measured build #10,$SAMPLE")
+        }
+
+        /**
+         * Asserts that this report file contains a single benchmarking scenario that uses a cold daemon.
+         */
+        void containsColdDaemonScenario(String gradleVersion, String name = "default", List<String> tasks) {
+            def lines = this.lines
+            assert lines.size() == totalLinesForExecutions(11)
+            assert lines.get(0) == "scenario,${name}"
+            assert lines.get(1) == "version,Gradle ${gradleVersion}"
+            assert lines.get(2) == "tasks,${tasks.join(", ")}"
+            assert lines.get(3) == "value,total execution time"
+            assert lines.get(4).matches("warm-up build #1,$SAMPLE")
+            assert lines.get(5).matches("measured build #1,$SAMPLE")
+            assert lines.get(6).matches("measured build #2,$SAMPLE")
+            assert lines.get(14).matches("measured build #10,$SAMPLE")
+        }
+
+        /**
+         * Asserts that this report file contains a single benchmarking scenario that uses no daemon.
+         */
+        void containsNoDaemonScenario(String gradleVersion, String name = "default", List<String> tasks) {
+            def lines = this.lines
+            assert lines.size() == totalLinesForExecutions(11)
+            assert lines.get(0) == "scenario,${name}"
+            assert lines.get(1) == "version,Gradle ${gradleVersion}"
+            assert lines.get(2) == "tasks,${tasks.join(", ")}"
+            assert lines.get(3) == "value,total execution time"
+            assert lines.get(4).matches("warm-up build #1,$SAMPLE")
+            assert lines.get(5).matches("measured build #1,$SAMPLE")
+            assert lines.get(6).matches("measured build #2,$SAMPLE")
+            assert lines.get(14).matches("measured build #10,$SAMPLE")
+        }
+    }
+
+    static class LogFile extends FileFixture {
+        LogFile(File logFile) {
+            super(logFile)
+        }
+
+        /**
+         * Asserts that this log file contains a single benchmarking scenario that uses a warm daemon.
+         */
+        void containsWarmDaemonScenario(String gradleVersion, String name = null, List<String> tasks) {
+            // Probe version, 6 warm up, 10 builds
+            containsScenario(name, gradleVersion)
+            assert find("* Running warm-up build").size() == 6
+            assert find("* Running measured build").size() == 10
+            assert find("<gradle-version: $gradleVersion>").size() == 17
+            assert find("<daemon: true").size() == 17
+            assert find("<daemon: false").size() == 0
+            assert find("<tasks: [:help]>").size() == 1
+            assert find("<tasks: [${tasks.join(", ")}]>").size() == 16
+            assert containsOne("<invocations: 16>")
+        }
+
+        /**
+         * Asserts that this log file contains a single benchmarking scenario that uses a cold daemon.
+         */
+        void containsColdDaemonScenario(String gradleVersion, String name = null, List<String> tasks) {
+            // Probe version, 1 warm up, 10 builds
+            containsScenario(name, gradleVersion)
+            assert find("* Running warm-up build").size() == 1
+            assert find("* Running measured build").size() == 10
+            assert find("<gradle-version: $gradleVersion>").size() == 12
+            assert find("<daemon: true").size() == 12
+            assert find("<daemon: false").size() == 0
+            assert find("<tasks: [:help]>").size() == 1
+            assert find("<tasks: [${tasks.join(", ")}]>").size() == 11
+            assert find("<invocations: 1>").size() == 12
+        }
+
+        /**
+         * Asserts that this log file contains a single benchmarking scenario that uses no daemon.
+         */
+        void containsNoDaemonScenario(String gradleVersion, String name = null, List<String> tasks) {
+            // Probe version, 1 warm up, 10 builds
+            containsScenario(name, gradleVersion)
+            assert find("* Running warm-up build").size() == 1
+            assert find("* Running measured build").size() == 10
+            assert find("<gradle-version: $gradleVersion>").size() == 12
+            assert find("<daemon: true").size() == 1
+            assert find("<daemon: false").size() == 11
+            assert find("<tasks: [:help]>").size() == 1
+            assert find("<tasks: [${tasks.join(", ")}]>").size() == 11
+            assert find("<invocations: 1>").size() == 12
+        }
+
+        private containsScenario(String name, String gradleVersion) {
+            def display = name == null ? "scenario" : "scenario ${name}"
+            assert containsOne("* Running $display using Gradle $gradleVersion (scenario 1/1)")
+        }
+    }
+
+    static int totalLinesForExecutions(int numberOfExecutions) {
+        return NUMBER_OF_HEADERS + numberOfExecutions
+    }
+}
